# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ShopSphere API",
        "description": "ShopSphere is a production-ready e-commerce platform. This API provides endpoints for product browsing, shopping cart management, wishlist, order processing, and user authentication.",
        "termsOfService": "",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.shopsphere.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/categories/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by slug",
                "description": "Returns a single category by its URL slug.",
                "operationId": "015a7b9d140957c1a14c67c5318977e2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "electronics"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List user orders",
                "description": "Returns a list of all orders placed by the authenticated user.",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Orders retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "description": "Returns detailed information about a specific order, including items and shipping.",
                "operationId": "88d27387b1e409e075c18609d4a0c300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/products/featured": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get featured products",
                "description": "Returns a list of featured products.",
                "operationId": "787711e5f2b08d01a9338d339746e9e1",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of products to return",
                        "schema": {
                            "type": "integer",
                            "default": 8
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of featured products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search products",
                "description": "Search for products by name, description, or other attributes.",
                "operationId": "6b588a52a56a41d175df187e5fb615c8",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Search results."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "description": "Pagination metadata for paginated responses",
                "properties": {
                    "current_page": {
                        "description": "Pagination metadata.",
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 200
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 20,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Category": {
                "description": "Product category",
                "properties": {
                    "id": {
                        "description": "Category schema.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "slug": {
                        "type": "string",
                        "example": "electronics"
                    },
                    "description": {
                        "type": "string",
                        "example": "All electronic products",
                        "nullable": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost:8000/storage/categories/category.jpg",
                        "nullable": true
                    },
                    "products_count": {
                        "type": "integer",
                        "example": 25
                    }
                },
                "type": "object"
            },
            "Product": {
                "description": "Product with details",
                "properties": {
                    "id": {
                        "description": "Product schema.",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Wireless Bluetooth Headphones"
                    },
                    "slug": {
                        "type": "string",
                        "example": "wireless-bluetooth-headphones"
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-001"
                    },
                    "description": {
                        "type": "string",
                        "example": "High-quality wireless headphones with noise cancellation"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "discount_price": {
                        "type": "number",
                        "format": "float",
                        "example": 79.99,
                        "nullable": true
                    },
                    "final_price": {
                        "type": "number",
                        "format": "float",
                        "example": 79.99
                    },
                    "has_discount": {
                        "type": "boolean",
                        "example": true
                    },
                    "discount_percent": {
                        "type": "integer",
                        "example": 20
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost:8000/storage/products/product.jpg"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "url": {
                                    "type": "string",
                                    "format": "uri"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "stock_quantity": {
                        "type": "integer",
                        "example": 50
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "average_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5,
                        "nullable": true
                    },
                    "reviews_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "OrderItem": {
                "description": "Item in an order",
                "properties": {
                    "id": {
                        "description": "Order Item schema.",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Wireless Bluetooth Headphones"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 199.98
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
            "Order": {
                "description": "Customer order",
                "properties": {
                    "id": {
                        "description": "Order schema.",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_number": {
                        "type": "string",
                        "example": "ORD-20240601-001"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "payment_method": {
                        "type": "string",
                        "enum": [
                            "cod"
                        ],
                        "example": "cod"
                    },
                    "payment_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "paid",
                            "failed"
                        ],
                        "example": "pending"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 199.98
                    },
                    "tax": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "shipping_fee": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 229.98
                    },
                    "shipping_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "shipping_phone": {
                        "type": "string",
                        "example": "+85512345678"
                    },
                    "shipping_address": {
                        "type": "string",
                        "example": "123 Street"
                    },
                    "shipping_city": {
                        "type": "string",
                        "example": "Phnom Penh"
                    },
                    "shipping_state": {
                        "type": "string",
                        "example": "Phnom Penh",
                        "nullable": true
                    },
                    "shipping_zip": {
                        "type": "string",
                        "example": "12000",
                        "nullable": true
                    },
                    "shipping_country": {
                        "type": "string",
                        "example": "Cambodia"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Leave at the door",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token. Get it from POST /api/login or /api/register.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}